# docs/swagger.yaml
basePath: /api
definitions:
  controllers.CallActionRequest:
    properties:
      action:
        type: string
      call_id:
        type: string
      reason:
        type: string
    required:
    - action
    - call_id
    type: object
...
  controllers.CallRequest:
    properties:
      device_id:
        type: string
      resident_id:
        type: string
    required:
    - device_id
    - resident_id
    type: object
  controllers.CreateAdminRequest:
    properties:
      email:
        example: admin@example.com
        type: string
      password:
        example: Admin@123
        type: string
      phone:
        example: "13800138000"
        type: string
      username:
        example: admin123
        type: string
    required:
    - email
    - password
    - phone
    - username
    type: object
...
  controllers.ErrorResponse:
    properties:
      code:
        example: 401
        type: integer
      data: {}
      message:
        example: Invalid username or password
        type: string
    type: object
...
  controllers.InitiateCallRequest:
    properties:
      callee_id:
        type: string
      caller_id:
        type: string
    required:
    - callee_id
    - caller_id
    type: object
...
  controllers.PublishDeviceStatusRequest:
    properties:
      battery:
        type: integer
      device_id:
        type: string
      online:
        type: boolean
      properties:
        additionalProperties: true
        type: object
    required:
    - device_id
    type: object
...
  controllers.UpdateAdminRequest:
    properties:
      email:
        example: admin@example.com
        type: string
      password:
        example: NewPassword@123
        type: string
      phone:
        example: "13800138000"
        type: string
    type: object
...
host: localhost:20033
info:
  contact:
    email: support@yourcompany.com
    name: API Support
    url: http://www.yourcompany.com/support
  description: A comprehensive door access management system with video calling capabilities
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ILock HTTP Service API
  version: "1.0"
paths:
  /admins:
    get:
      consumes:
      - application/json
      description: Get a list of all admin users with pagination
      parameters:
      - description: Page number, default is 1
        in: query
        name: page
        type: integer
      - description: Items per page, default is 10
        in: query
        name: page_size
        type: integer
      - description: Search keyword for username, phone, etc.
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Admin List
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Create a new admin user
      parameters:
      - description: Admin information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateAdminRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Admin
      tags:
      - Admin
  /admins/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an admin user with the specified ID
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Admin
      tags:
      - Admin
    get:
      consumes:
      - application/json
      description: Get details of a specific admin by ID
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Admin By ID
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update an existing admin user
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated admin information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Admin
      tags:
      - Admin
  /api/mqtt/callee-action:
    post:
      consumes:
      - application/json
      description: Process callee's reject, hangup or timeout actions
      parameters:
      - description: Call action request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CallActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Handle MQTT Callee Action
      tags:
      - MQTT Call
  /api/mqtt/caller-action:
    post:
      consumes:
      - application/json
      description: Process caller's cancel or hangup actions
      parameters:
      - description: Call action request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CallActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Handle MQTT Caller Action
      tags:
      - MQTT Call
  /api/mqtt/device/status:
    post:
      consumes:
      - application/json
      description: Publish device status information via MQTT
      parameters:
      - description: Device status information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.PublishDeviceStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Publish Device Status
      tags:
      - MQTT
  /api/mqtt/initiate:
    post:
      consumes:
      - application/json
      description: Start a new call via MQTT
      parameters:
      - description: Call initiation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.InitiateCallRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Initiate MQTT Call
      tags:
      - MQTT Call
  /api/mqtt/system/message:
    post:
      consumes:
      - application/json
      description: Publish system message via MQTT
      parameters:
      - description: System message information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.PublishSystemMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Publish System Message
      tags:
      - MQTT
  /api/rtc/call:
    post:
      consumes:
      - application/json
      description: Initiate a video call between a device and a resident
      parameters:
      - description: Call request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CallRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Start Video Call
      tags:
      - RTC
  /api/rtc/token:
    post:
      consumes:
      - application/json
      description: Get a RTC token for real-time communication
      parameters:
      - description: Token request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.TokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get RTC Token
      tags:
      - RTC
  /auth/login:
    post:
      consumes:
      - application/json
      description: Process user login and return JWT token with different permissions
        based on user role
      parameters:
      - description: Login request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response with token
          schema:
            allOf:
            - $ref: '#/definitions/controllers.LoginResponse'
            - properties:
                data:
                  $ref: '#/definitions/controllers.LoginData'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: User Login
      tags:
      - Auth
...
  /devices:
    get:
      consumes:
      - application/json
      description: 获取所有设备的列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Device'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取所有设备
      tags:
      - device
    post:
      consumes:
      - application/json
      description: 创建一个新的门禁设备
      parameters:
      - description: 设备信息
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/controllers.DeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建新设备
      tags:
      - device
  /devices/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除设备
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除设备
      tags:
      - device
    get:
      consumes:
      - application/json
      description: 根据ID获取设备信息
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取单个设备
      tags:
      - device
    put:
      consumes:
      - application/json
      description: 根据ID更新设备信息
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      - description: 设备信息
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/controllers.DeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新设备信息
      tags:
      - device
  /devices/{id}/status:
    get:
      consumes:
      - application/json
      description: Get the current status of a device by ID
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Device Status
      tags:
      - device
  /emergency/alarm:
    post:
      consumes:
      - application/json
      description: Trigger an emergency alarm and notify relevant personnel
      parameters:
      - description: Alarm request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.EmergencyAlarmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Trigger Emergency Alarm
      tags:
      - Emergency
  /emergency/contacts:
    get:
      consumes:
      - application/json
      description: Get a list of all emergency contacts in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Emergency Contacts
      tags:
      - Emergency
  /emergency/notify-all:
    post:
      consumes:
      - application/json
      description: Send notification to all users during an emergency
      parameters:
      - description: Notification request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.EmergencyNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Notify All Users
      tags:
      - Emergency
  /emergency/unlock-all:
    post:
      consumes:
      - application/json
      description: Unlock all doors in the system during an emergency
      parameters:
      - description: Unlock request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.EmergencyUnlockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Emergency Unlock All Doors
      tags:
      - Emergency
  /residents:
    get:
      consumes:
      - application/json
      description: Get a list of all residents in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Resident List
      tags:
      - Resident
    post:
      consumes:
      - application/json
      description: Create a new resident account, requiring association with a specific
        device
      parameters:
      - description: Resident information - name, phone and device ID are required,
          email is optional
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.ResidentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success response with created resident details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request, device not found or phone number already in use
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Resident
      tags:
      - Resident
  /residents/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a resident with the specified ID
      parameters:
      - description: Resident ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Resident
      tags:
      - Resident
    get:
      consumes:
      - application/json
      description: Get details of a specific resident by ID
      parameters:
      - description: Resident ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Resident By ID
      tags:
      - Resident
    put:
      consumes:
      - application/json
      description: Update details of a resident with the specified ID
      parameters:
      - description: Resident ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated resident information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateResidentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Resident
      tags:
      - Resident
  /staffs:
    get:
      consumes:
      - application/json
      description: Get a list of all property staff members, with pagination and search
        support
      parameters:
      - description: Page number, default is 1
        in: query
        name: page
        type: integer
      - description: Items per page, default is 10
        in: query
        name: page_size
        type: integer
      - description: Search keyword for name, phone, etc.
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Property Staff List
      tags:
      - Staff
    post:
      consumes:
      - application/json
      description: Create a new property staff member
      parameters:
      - description: Property staff information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateStaffRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Property Staff
      tags:
      - Staff
...
  /trtc/call:
    post:
      consumes:
      - application/json
      description: Initiate a Tencent Cloud video call between a device and a resident
      parameters:
      - description: Call request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.TencentCallRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Start Tencent Video Call
      tags:
      - TencentRTC
...
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer: ` prefix'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"